+++ lib/lazycouchbase/ui/status_bar.rb
@@ -9,7 +9,8 @@ module Lazycouchbase
     class StatusBar
       HINTS = {
         normal: "j/k: move │ enter: open │ /: filter │ :: query │ ?: help │ q: quit",
-        query: "enter: run │ ctrl-e: explain │ ↑/↓: history │ esc: back",
+        query: "enter: run │ ctrl-e: explain │ tab: snippets │ ↑/↓: history │ esc: back",
+        snippet: "type to narrow │ ↑/↓: move │ enter: insert │ ctrl-o: docs │ esc: back",
         document: "j/k: move │ /: search │ t: keys │ y/Y: yank │ esc: back",
         document_search: "type to search │ enter: jump │ esc: cancel",
         help: "esc: close",